+++ rastervision/backend/torch_utils/chip_classification/folder.py
@@ -6,6 +6,7 @@
 from torchvision.datasets.vision import VisionDataset
 
 from PIL import Image
+from numpy import (array,ndarray)
 
 import os
 import os.path
@@ -155,7 +156,16 @@ class DatasetFolder(VisionDataset):
         path, target = self.samples[index]
         sample = self.loader(path)
         if self.transform is not None:
-            sample = self.transform(sample)
+            # Convert PIL image to numpy array first, because albumentations used for
+            # the transforms expect an array.
+
+            sample_np = array(sample)
+            sample = self.transform(image=sample_np)
+            if isinstance(sample,ndarray):
+                sample = Image.fromarray(sample['image'])
+            sample = sample['image']
+            # Otherwise it is a tensor and we should not change its type
+
         if self.target_transform is not None:
             target = self.target_transform(target)
 
@@ -171,11 +181,11 @@ IMG_EXTENSIONS = ('.jpg', '.jpeg', '.png', '.ppm', '.bmp', '.pgm', '.tif',
 
 def pil_loader(path):
     # open path as file to avoid ResourceWarning (https://github.com/python-pillow/Pillow/issues/835)
+    from numpy import asarray
     with open(path, 'rb') as f:
         img = Image.open(f)
         return img.convert('RGB')
-
-
+        
 def accimage_loader(path):
     import accimage
     try:
@@ -183,7 +193,6 @@ def accimage_loader(path):
     except IOError:
         # Potentially a decoding problem, fall back to PIL.Image
         return pil_loader(path)
-
 
 def default_loader(path):
     from torchvision import get_image_backend